Introduce new tile locking mechanism and minor cosmetic changes
This set of changes enables GeglTiles to lock for both read and
write operations, allowing the tile to synchronize GPU data to/from
the data in system memory.  This also contains some minor changes
to most existing files that currently lock the tile when writing. Note
that the latter changes are made to only satisfy the compiler. Support
for read locks will be added to the relevant files in later commits.

Re: cosmetic changes, nothing much really; a few code-style
formatting here and there.  And oh, apparently, I forgot to commit the
implementations for both gegl_tile_get_gpu_data() and _get_data()
in previous commits.  This commit contains both implementations.