Change gegl_class_register_alternate_vfunc() to support registering GPU operation processors
When an operation processor whose name begins with the string "gpu"
is registered through this method, the processor is added to a separate
dispatch data made specifically for GPU operation processors.

This was made so that GPU implementations can support best, good,
and bad variants in addition to their reference implementations.