GeglOperationPointFilter: Implement GPU processing of pixels
When both GPU acceleration is enabled and at least one GPU processor
is available for a point filter, GEGL will make use of the GPU processor
and process GPU textures instead of pixels in main memory.

Refer to the definition of GeglOperationPointFilterGpuProcessor in
gegl/gegl/operation/gegl-operation-point-filter.h to see how a point filter
GPU processor should be prototyped.